[Morello] Add a pass to change the way we are creating addresses for aliases in purecap.
The motivation is that the global merge pass will replace
globals with aliases to a merged global.

The merged global address is produced by loading a capability
from a pc-relative relocation (same as before).

Safe loads/stores can store to the merged global directly,
without the need to further refine bounds.

Unsafe uses need to refine the bounds of the merged global.