[lld] Merge getMorelloTargetSize into getTargetSize
This ensures Morello gets the various bug fixes and improvements made to
the normal CHERI getTargetSize since it made its own copy. As part of
this, fix the hacky handling of symbol sizes added for Morello as
follows.

getSize only respected the isSectionStartSymbol field for CheriABI, but
hybrid Morello binaries can have capabilities too, so add a new boolean
parameter for CheriCapRelocsSection to use in order to force it to be
honoured.

Linker scripts were also not setting isSectionStartSymbol for things
that looked like section start symbols, so getSize would never use the
section's bounds for them.

Because of these two issues, getMorelloTargetSize overrode a value of
0 for symbols that had a name that looked like a section start symbol
rather than taking into account the additional restrictions from Writer
on whether a symbol is a section start symbol. By ensuring the symbol
information is right in the first place this hack is no longer needed.

Note: isSectionStartSymbol is both a field in Defined that encompasses
all the checks, and a function that just matches the name. The former is
the property that should be used to decide whether to use the section
size, and the latter is used both as part of determining the value for
that field and for suppressing warnings about zero-length capabilities
(as would otherwise seen for empty sections).

This fixes the following assertion previously triggered by the new test
(copied from the existing CHERI capreloc-aether test):

  Assertion failed: (OffsetInOS <= OS->size), function getMorelloTargetSize, file lld/ELF/Arch/Cheri.cpp, line 739.