lld support for R_MORELLO_DESC_ADR_PREL_PG_HI20
Switches the instruction to adrdp if the target symbol (S) is in a .descdata
section. In this case the relocation operation = Page(S+A)-Page(D), where
D = Address of the .descdata section

Switches the instruction to adrp if the symbol is not in a .descdata section.
In this case the relocation operation is same as R_MORELLO_ADR_PREL_PG_HI20.

Change-Id: I46770b47596238be73b65f567e8b7ce9e625388c