[lld][morello] Add --strip-note-cheri option
Strips the .note.cheri section from the final image. Because the Alloc bit
is set, the presence of this section affects CHERI concentrate alignment
and results in a different image layout.

Now, for images with the e_flag EF_AARCH64_CHERI_PURECAP set, the
--morello-c64-plt linker option can be omitted and stripping the .note.cheri
section will result in an identical image layout.

Change-Id: I27587ab1bfecc9159947a1808f9ff6133052df1f