[lld][Morello][descriptor-abi] Fix .got.plt relro/non-relro behavior
The previous implementation added the .got.plt to the RO region of the
PT_MORELLO_DESC segment unconditionally. However the -z now linker switch
controls if the .got.plt is RELRO or not.
Therefore use the same conditions to include the .got.plt in the
PT_MORELLO_DESC RO region and add a check for the -z now case.

Change-Id: I520ff8251d4e92a0fe298ba6391271be3f8df73a