[CHERIseed] Remove AtomicBool and use atomic_uint8_t instead
The extra abstraction introduced by AtomicBool can be confusing due to
the overloaded = operator. Use atomic_uint8_t directly with the helper
functions: atomic_load_relaxed and atomic_store_relaxed. Although this
makes the code more verbose, it is easier to understand. Besides, this
is more in line with what other sanitizers do.

Change-Id: I0fc6193f66ac09066ccfd98af0ff93b483def2a5