[CHERIseed] Implement runtime controls for CHERI semantics
This patch introduces a new RT API for enabling/disabling checks. It
allows the user to supply a 64-bit mask that indicates which checks to
turn ON/OFF. The lower 32-bits represent permission checks, as per the
CHERI_PERM_* macros defined in cheriintrin.h, and the upper 32-bits
represent non-permission checks, such as tag, bounds and alignment
checks. By default, all checks are enabled.

Change-Id: I5c8c19b4a18e4e79c46456e618878570b35ef998