[morello][lld][ELF] Omit RELATIVE relocs for hidden weak symbols in PI executables
Fix for Morello llvm-project Issue #50

RELATIVE relocations against hidden weak symbols in -pie binaries are
unnecessary because the result is going to be zeros in the address part
of the fragment which are already set by the static linker. Also the
runtime could process these RELATIVE relocations and set a non-zero
address which can result in incorrect behaviour.

Fix by omitting RELATIVE relocations for hidden weak symbols in pie.