[CHERIseed] Refactor accessors
There is no need for accessors now that global inits have
been refactored. The reason of this refactor is that when
compiling libcxxabi there was another corner case which
would have been hard to support.

To decrease the complexity of the code, this refactor splits
up how GlobalVariables and GlobalAliases are handled:
  1. GlobalVariables are mapped, but not their initializers.
  2. GlobalAliases are mapped, but not their initializers.
  3. Initializers of GlobalVariables are processed.
  4. Alisees of GlobalAliases are processed.
  5. Initializers are emitted.

This way there is no need to defer resolution (where the issue was).
The code is simpler and there are no more accessor calls.

With this change there are no regressions and libcxxabi builds.

Change-Id: I41041113cdb53535e8e507cf8467553fc3a63094