[CHERIseed] Add more instrumentation to data stores
When a check_access has store semantics too, the tags
related to that access are locked. This is required
so that no racing tag writes can create aribtrary random
capabilitites. The end of such an instrumented store is
marked with check_access_end and it makes sure that tags
related to that data store access are unlocked.

Co-authored-by: Tamas Petz <tamas.petz@arm.com>
Co-authored-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>

Change-Id: Ie841ccb338f3a6301ef5f80ae9077a297908fae0