[CHERIseed] Change how PHI nodes are copied
There were issues with PHI nodes in several loop layout.

This change moves the PHI copies to the end of the
predecessor blocks, ensuring that their values are not
used before the PHI node takes their values. These are
called "tail" copies.

However, making one copy in cases where a PHI node uses
another PHI node is not enough. In these cases there is
a further need for another copy after the PHI nodes. These
are called "head" copies.

The fundamental issue to solve is the fact that without
copies, pointers to capability storage locations start
to alias one another.

Change-Id: I5632efb5ac9c105b88bcc3a822fd14e40ca6483c