[Morello] Rework morello-spill-pair.ll to pass post-merge
Unoptimised code generation no longer decides to use a pair and instead
splits it. Disabling -O0 however both stops performing a spill and
lowers the atomicrmw xchg to a ldxp/stxp loop.

Rewrite this test so it performs an actual cmpxchg to ensure it gets a
casp (the only instruction that seems to require a consecutive pair of
registers and thus is allocated to a register pair class), disable -O0
so the registers aren't split and use inline assembly to clobber all
GPRs and force the pair to be spilled.