[DRAFT][Morello][LLD][ELF] Improve information provided by R_MORELLO_JUMP_SLOT relocations
The previous implementation of R_MORELLO_JUMP_SLOT in the static linker only
provided the address to the PLT header. The limit, size parts of the fragment
and the addend of the R_MORELLO_JUMP_SLOT relocation was unused.

https://github.com/ARM-software/abi-aa/pull/249 Updates the specification of
R_MORELLO_JUMP_SLOT to provide information similar to R_MORELLO_RELATIVE.

This means the under the new static linker implementation
- 'address' part of the fragment = 'capability aligned base address' for executable capabilities.
- 'size' part of the fragment = 'capability aligned limit' for executable capabilities.
- 'permissions' part of the fragment = 'executable' (0x04).
- 'addend' of R_MORELLO_JUMP_SLOT will be such that, 'addend' + 'address' = address of PLT[0]