[Morello] Deprecate .capinit directive
For over 3.5 years .chericap has been the preferred form in Morello LLVM
over the hacky Morello-specific .capinit. The latter is error-prone due
to not allocating space for the capability, instead being similar to
.reloc. Formally deprecate the directive so that eventually it can be
removed, and so that in the meantime people don't risk using it without
being aware that it's not the same as .chericap.