[ELF] Fix mismerge in addCapDynRelocsSymbols
82a5fca8d986 ("[ELF] Remove unneeded binding parameter from
addOptionalRegular. NFC") failed to remove the extra argument from the
call to addOptionalRegular, causing STB_WEAK to be implicitly cast to a
bool for canBeSectionStart. As it happens, the bool defaults to true and
STB_WEAK is non-zero, so this was harmless, but wrong and confusing.

Whilst here, rename the function and reformat the code to align with
addRelIpltSymbols.