[Morello] Handle explicit addend in global capabilities
This should probably be fixed more cleanly by marking the symbol table
entries with the LSB (e.g. by adding a flag to MCSymbolELF) but for now
handling a binary expression in AsmPrinter.cpp allows compiling again
after the upstream merge. While fixing this also update the
morello-code-capabilities.ll test to be a bit more detailed.