[Morello] Remove downstream hack breaking register pairs
The assembly parser now appears to handle these correctly so we no
longer need this hack. This incorrect definition also completely broke
existing RegMask value (this was noticed because it causes e.g. x0 to be
included in the Darwin AAPCS regmask and that breaks the tail call
analysis). This fixes lots of tests.