[Morello][ELF] Fix mismerge that caused incorrect capability relocs
The relocateAlloc function was moved to the target specific files, but
I failed to notice that the control flow changed to call relocate()
after the switch statement. We have to continue the loop after writing
the capability fragment data since otherwise the subsequent relocate()
call will overwrite it with incorrect values.