[ELF][Morello] Add a mutex to fix race condition
I was seeing assertion errors when accessing relocsMap with a backtrace
pointing to `MorelloTLSLEDataSection::addTLSLEData` while building
CheriBSD. Building LLVM with TSan showed the race condition while
accessing relocsMaps. Fix this by adding a mutex. The other functions
that access relocsMap are currently always single-threaded after the
initial relocation scanning so there is no need for a lock there.

This did not cause problems prior to the LLVM-17 merge since the code
was previously single-threaded.