[ELF][Morello] Implement new R_MORELLO_JUMP_SLOT ABI
Rather than copy the ABI for R_AARCH64_JUMP_SLOT, i.e. use an addend of
0 and put the VA of &.plt[0] in the GOT entry, which doesn't give the
loader any capability information, the new ABI, as documented in
aaelf64-morello, encodes a capability fragment for &.plt[0], in the same
manner as R_MORELLO_RELATIVE, except with a non-zero symbol index,
corresponding to the function the PLT entry is for.