[NFC][ELF][Morello] Use TargetInfo::writeIgotPlt rather than relocateAlloc
The current code is confusing, as it unconditionally adds a fragment for
the symbol, even though that makes no sense for a normal .got.plt entry
which is against a preemptible symbol. This only worked because it was a
no-op for GotPltSection, as its writeTo does not call relocateAlloc, and
so only IgotPltSection actually did anything for this fragment.

Instead, follow how other architectures work and write the fragment
(akin to the addend on other architectures when using REL) within
writeIgotPlt.