[ELF][Morello] Don't emit fragment for relocations with symbols
Emitting a fragment for a symbol that is preemptible (and may not be
defined) does not make sense, as the symbol must be looked up at run
time to get the right metadata; for undefined symbols it's particularly
useless to use an arbitrary value of 2, i.e. read write with 0 size, for
the second half of the fragment.

Instead, leave it as 0, and refactor the code to make it clearer what's
going on. Note that we now no longer need separate RelExprs for exec and
non-exec, since in all cases we have the actual Defined that was used to
add the relocation and thus can just query whether it's a function (or
IFUNC).