[NFCI][ELF][Morello] Implicitly set localCapRelocsMode for dynamic objects
Now that we properly implement CapRelocsMode::ElfReloc for Morello we
should just use that for the dynamic case rather than leave it
pretending that we're using legacy caprelocs when we're not.

Note that the getTargetSize condition should have been hasDynSymTab as
it was in addMorelloRelativeRelocation; otherwise this will read 8 bytes
past the end of the fragment if producing a dynamically-linked
executable, which is often, but not always, 0, and may even be out of
bounds if at the end of the data section.