[ELF][Morello] Fix and clean up relative relocation handling
Firstly, R_MORELLO_IRELATIVE never takes this code path, so constant
fold that case and assert that we only see the expected (original)
relocations. Whilst here also assert that we're in a case where we
aren't using cap relocs.

In terms of picking between in.relaDyn and mainPart->relaDyn, this
should be tied to whether we added mainPart->relaDyn or not, which is
based on whether we have a dynamic symbol table; currently relocations
can end up in a less-optimal order if you enable --local-caprelocs=elf
for a dynamic binary (namely, relative relocations to symbols not
exported end up not being first, and so DT_RELACOUNT does not include
them).