[NFC][ELF][Morello] Don't use addend in getMorelloBaseAddress
The single use of this, via addMorelloCapabilityFragment creating an
R_MORELLO_CAPFRAG_BASE relocation, ends up with an addend of 0 (even if
this is for a capability relocation with an addend), so this currently
works as expected, but this is a footgun that should be removed, and
will make an upcoming change clearer. Fix this API to be "the base to
put in a fragment for a RELATIVE-like relocation against sym+a".