[ELF][Morello] Reimplement fragment addend calculation as a RelExpr
DynamicReloc::Kind is overly complex upstream and should have just been
a bool (or equivalent) indicating whether the reference the symbol in
the output, with the addend calculation chosen solely by the given
RelExpr. We should avoid making this worse downstream and instead
simplify our implementation by following this principle.

Note that addCapDynamicRelocation's use of AgainstSymbol was bogus,
likely as a lazy way to allow sym to be provided whilst still getting
the original addend. It should really either have been a separate case
with AddendOnly and no symbol or AddendOnlyWithTargetVA and R_ADDEND.
This doesn't seem to have mattered much in practice though, but this
does resolve that historic mistake.