[ELF][AArch64] Remove special case for in.relaDyn creation
When --pack-dyn-relocs=android and --local-caprelocs=elf are both passed
for AArch64, if we have a dynamic symbol table, each partition's
.rel[a].dyn will be of type SHT_ANDROID_RELA, but this extra section
will be of type SHT_RELA, leading to a section type mismatch error.
There's no reason that I can see for this special case to exist when it
doesn't for the normal partition sections.

Note also this means we honour -z combreloc (and its default value of
true) for this section, again just as is already done for the normal
partition's section.