[NFC][ELF][Morello] Drop need for writeFragment* methods
Firstly, writeFragmentAddress is what reloc will do normally. Secondly,
writeFragmentSizeAndPermissions is the same, except for the fact that it
reorders the bytes (at least on a little-endian host; I suspect it might
not be correct on a big-endian one). If we get the bytes round the right
way in the first place then we can also just use relocate for this.