[ELF][Morello] Use correct section offset for getTargetSize for ELF relocs
The offset getTargetSize wants is for the whole capability, not whatever
internal offset we happen to have for the split static relocations. This
avoids needing to infer the offset for the compiler-provided size, as
all it's really trying to do here is infer which function called it. It
also means that any diagnostics print the correct offset.