[NFCI][ELF][Morello] Rework relative*Rel to align more with CHERI LLD
CHERI LLD assumes that relative(Func)Rel are for integers, just as
symbolicRel is for integers. Align with that rather than hacking around
the difference in addRelativeReloc.

Also take advantage of relativeCapFuncRel only being set for Morello
when cheriEmitCodePtrRelocs is true in addRelativeCapabilityRelocation,
as this generalises this part of the code to future non-Morello use.

Finally, RelocationBaseSection::partitionRels can take advantage of
these various members to avoid needing to special-case AArch64 so much;
now only the special descriptor ABI relocations remain.