[NFC][ELF][Morello] Simplify addMorelloRelativeElfReloc
The only reason we don't modify dynType is so we can later assert that
it was a sensible value. Hoist the assert instead and then just modify
the input. Note that realDynType = R_MORELLO_RELATIVE was a no-op as it
was guarded by dynType == R_MORELLO_RELATIVE and realDynType was already
assigned to dynType's value.

Hoist a couple of other asserts whilst here to facilitate a future
refactor.