[NFC][ELF][Morello] Handle fragments for relative relocations via addReloc
Logically we have a capability addend here (an encoding, albeit as a
fragment not the actual raw bits, of the capability we wish to have at
run time, without the load base address applied). By pushing the code
into addRelocCap and teaching addReloc we can tidy up the interface to
look like how non-CHERI relative relocations work, hiding the details of
fragment emission.