Duplicate merge sections for each compartment
For each SHF_MERGE section not assigned to an explicit compartment,
leave the original section in the default compartment but create
duplicate sections with the same content for each compartment.  For
each symbol defined in the original merge section, a duplicate symbol
is created for each compartment's copy.

InputFile::getRelocTargetSym now accepts a compartment reference.  If
the symbol index names a symbol from a merge section, the
corresponding symbol from the compartment's copy of the merge section
is returned.

Note that without --gc-sections, each compartment is allocated an
entire copy of each merge section, even if a compartment does not use
any symbols from that section.