[ELF] Allow --local-caprelocs=elf for EM_ARM
CheriBSD's ld.lld wrapper passes various flags by default to ensure it
gets the right ABI even if the compiler driver is bypassed. However,
unlike cc where the target must be given, there is often no emulation
given to ld.lld, as it can be inferred, and so this wrapper does not
know whether it is linking an EM_AARCH64 or an EM_ARM binary, and thus
has no way of only passing --local-caprelocs=elf for the former. It's
therefore convenient to treat EM_ARM like EM_AARCH64 here and allow the
flag to be passed even though it does nothing. In theory this is also
helpful for anyone passing explicit -Wl,--local-caprelocs=elf to cc
without needing to worry about --target and -m32, but that is not the
primary motivation.