[test][lldb][Morello] Avoid using descriptor ABI to imply purecap
This test doesn't care about the descriptor ABI, it just wants to create
a purecap target, but there's no command to force a given ABI outside of
the hacky Morello descriptor ABI one. Switch this test to using a
barebones purecap object file instead from yaml2obj so it will continue
to work when descriptor ABI support is removed.