[ELF][Morello] Make Morello's __cap_relocs read-only
The generic CHERI section has been read-only for relative-style
caprelocs since 9276b97a4c17 ("Make __cap_relocs read-only if
RelativeCapRelocsOnly is true"), which never got mirrored to the Morello
implementation, nor binutils. Matching the abandoned Morello binutils is
not important, but aligning with generic CHERI is.

Note we have to move __cap_relocs in the linker script for one of the
tests (so it doesn't sit at the end of what was the old PCC range)
since, now that it's read-only, it's a candidate for being included in
the PCC range, and in this case it ends up being the last such section
and thus receives tail padding, rendering it invalid. This is also true
if you put .rela.dyn in such a location, but normally people don't do
that, so has gone unnoticed. This will be fixed by pulling in a smarter
PCC representability implementation from CHERI LLVM soon, but should not
pose a problem in the real world in the meantime.