[ELF][Morello] Fold MorelloCapRelocsSection into CheriCapRelocsSection
The implementations have converged over the years and with the latest
permissions abstraction it's relatively straightforward to merge the
two. This also uses the new getTargetType for determining how to encode
fragment permissions so there is a single source of truth for which type
to use across CHERI caprelocs, Morello caprelocs and Morello fragments.