[ELF][Morello] Stop using pointless MorelloTLSLEDataSection
MorelloTLSLEDataSection entries are entirely equivalent to IE GOT
entries, just with no associated dynamic relocation, because both 64-bit
words are known link-time constants. But we don't need a new section for
that, we can just disable the Local-Exec relaxations, and leave the
existing code in addTpOffsetGotEntry to recognise both are constants.

Note that morello-ld128-got-lo12-nc.s changes because this section had
16 byte alignment and was merged into .rodata, and this alignment forced
.rodata's alignment even if MorelloTLSLEDataSection was empty.